Check that the data is equispaced
The previous condition was *always* satisfied by the construction of inf and sup, therefore resulting in the application of equispaced quadrature rules to non-equispaced data. For the TRAPZ method, the default, the bug was not very harmful because integrate.simpson was always computing the non-equispaced trapezoid rule (at expenses of computational efficiency).